# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Product APIs TP Java por los pibes.",
    "version" : "1.0.0",
    "title" : "ProductStore APIs",
    "contact" : {
      "name" : "JEG",
      "url" : "https://www.facebok.com",
      "email" : "vigliancoj@gmail.com"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/productstore-back-1.0-SNAPSHOT/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/roles" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve todos los roles",
        "description" : "",
        "operationId" : "getRoles",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Roles encontrados",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Rol"
              }
            }
          },
          "204" : {
            "description" : "No se encontro el rol"
          }
        }
      },
      "post" : {
        "tags" : [ "Product" ],
        "summary" : "Crea un rol en formato JSON",
        "description" : "",
        "operationId" : "createRol",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "El rol fue creado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Se creo el rol."
          },
          "415" : {
            "description" : "No es formato JSON."
          }
        }
      },
      "put" : {
        "tags" : [ "Product" ],
        "summary" : "Actualiza un rol",
        "description" : "",
        "operationId" : "updateRol",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "El rol fue creado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Rol"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Rol actualizado"
          },
          "500" : {
            "description" : "Rol no encontrado"
          }
        }
      }
    },
    "/roles/count" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve el total de la cantidad de roles",
        "description" : "",
        "operationId" : "countRoles",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "Total de roles encontrados",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "204" : {
            "description" : "No existen roles"
          }
        }
      }
    },
    "/roles/{id}" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve un unico rol",
        "description" : "",
        "operationId" : "getRol",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Rol Encontrado",
            "schema" : {
              "$ref" : "#/definitions/Rol"
            }
          },
          "400" : {
            "description" : "Ingreso invalido. Debe poner un id mayor a 0."
          },
          "404" : {
            "description" : "Rol no encontrado."
          }
        }
      },
      "delete" : {
        "tags" : [ "Product" ],
        "summary" : "Elimina un rol a partir del id",
        "description" : "",
        "operationId" : "deleteRol",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Rol eliminado"
          },
          "400" : {
            "description" : "Ingreso invalido. Debe poner un id mayor a 0."
          },
          "500" : {
            "description" : "Rol no encontrado"
          }
        }
      }
    },
    "/usuarios" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve todos los usuarios",
        "description" : "",
        "operationId" : "getUsuarios",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Usuarios encontrados",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Usuario"
              }
            }
          },
          "204" : {
            "description" : "No se encontro el usuario"
          }
        }
      },
      "post" : {
        "tags" : [ "Product" ],
        "summary" : "Crea un usuario en formato JSON",
        "description" : "",
        "operationId" : "createUsuario",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "El usuario fue creado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Usuario"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Se creo el usuario."
          },
          "415" : {
            "description" : "No es formato JSON."
          }
        }
      },
      "put" : {
        "tags" : [ "Product" ],
        "summary" : "Actualiza un usuario",
        "description" : "",
        "operationId" : "updateUsuario",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "El usuario fue creado",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Usuario"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Usuario actualizado"
          },
          "500" : {
            "description" : "Usuario no encontrado"
          }
        }
      }
    },
    "/usuarios/count" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve el total de la cantidad de usuarios",
        "description" : "",
        "operationId" : "countUsuarios",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "Total de usuarios encontrados",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "204" : {
            "description" : "No existen usuarios"
          }
        }
      }
    },
    "/usuarios/{id}" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Devuelve un unico usuario",
        "description" : "",
        "operationId" : "getUsuario",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Usuario Encontrado",
            "schema" : {
              "$ref" : "#/definitions/Usuario"
            }
          },
          "400" : {
            "description" : "Ingreso invalido. Debe poner un id mayor a 0."
          },
          "404" : {
            "description" : "Usuario no encontrado."
          }
        }
      },
      "delete" : {
        "tags" : [ "Product" ],
        "summary" : "Elimina un usuario a partir del id",
        "description" : "",
        "operationId" : "deleteUsuario",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Usuario eliminado"
          },
          "400" : {
            "description" : "Ingreso invalido. Debe poner un id mayor a 0."
          },
          "500" : {
            "description" : "Usuario no encontrado"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Rol" : {
      "type" : "object",
      "properties" : {
        "idRol" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "nombreRol" : {
          "type" : "string",
          "description" : "Nombre del Rol"
        }
      },
      "description" : "Representacion de Rol"
    },
    "Usuario" : {
      "type" : "object",
      "properties" : {
        "idUsuario" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "nombreUsuario" : {
          "type" : "string",
          "description" : "Nombre del Usuario"
        },
        "apellidoUsuario" : {
          "type" : "string",
          "description" : "Apellido del Usuario"
        },
        "email" : {
          "type" : "string",
          "description" : "Email del Usuario"
        },
        "idRol" : {
          "description" : "Rol del Usuario",
          "$ref" : "#/definitions/Rol"
        },
        "username" : {
          "type" : "string",
          "description" : "Nombre de Usuario"
        },
        "password" : {
          "type" : "string",
          "description" : "Password de Usuario"
        }
      },
      "description" : "Representacion de Usuario"
    }
  }
}